Add extensible type system
This updates comes with a refactoring of the type system, along with
some internal fixes. The new type system makes it possible for users to
(re)define the types.

The type system introduces the concept of "Complex Types", which are PHP
classes that are serialized into a valid SQL value/type before being
sent to the database and then unserialized (also restored into the orig-
inal format) when retrieved.

This makes it possible to use feature rich classes like PHP's "DateTime"
in order to represent a TIMESTAMP column in SQL.

The internals of the \SQL\Record class have been partly refarctored in
order to accomodate the new type system.