Add "DEFAULT" and "NOT NULL" constraints
This commit introduces the ability to define default values for columns
and add "NOT NULL" constraints.

All columns have the "NOT NULL" constraint enabled by default. To dis-
able it, one must make the variable type nullable:

    public ?string $content;

To define default values, set default values in the class definition:

    public ?string $content = "\"sample text\"";

For now, default values for strings must have an entra pair of
parenthesis around them. But I'll get around to fixing that.